PXNewResultUtil
