+++ MercadoPagoSDK/MercadoPagoSDK/UI/PXResult/PXResultViewModel.swift
@@ -267,13 +267,9 @@ extension PXResultViewModel {
 
     //Receipt View
     func buildReceiptView() -> UIView? {
-        guard let props = getReceiptComponentProps(), let title = props.receiptDescriptionString else {
+        guard let data = PXNewResultUtil.getDataForReceiptView(paymentId: paymentResult.paymentId) else {
             return nil
         }
-        let attributedTitle = NSAttributedString(string: title, attributes: PXNewCustomView.titleAttributes)
-        let subtitle = props.dateLabelString ?? ""
-        let attributedSubtitle = NSAttributedString(string: subtitle, attributes: PXNewCustomView.subtitleAttributes)
-        let data = PXNewCustomViewData(firstString: attributedTitle, secondString: attributedSubtitle, thirdString: nil, icon: nil, iconURL: nil, action: nil, color: nil)
         let view = PXNewCustomView(data: data)
         return view
     }
@@ -303,7 +299,7 @@ extension PXResultViewModel {
 
     //Payment Method View
     func buildPaymentMethodView(paymentData: PXPaymentData) -> UIView? {
-        guard let data = PXNewCustomViewData.getDataFromPaymentData(paymentData, amountHelper: amountHelper) else {return nil}
+        guard let data = PXNewResultUtil.getDataForPaymentMethodView(paymentData: paymentData, amountHelper: amountHelper) else {return nil}
         let view = PXNewCustomView(data: data)
         return view
     }
